Add files via upload
Wind energy is clean, inexhaustible, inexpensive and widely distributed [1]. However, the randomness and intermittency of wind lead to uncertainty in wind power generation, which brings challenges to the corresponding energy management system and affects the reliability of the entire power grid. Therefore, accurate estimates of wind power curves ,which show the non-linear relationship between wind speed and wind power, are required for effective integration of wind power into the power system, as well as wind turbine condition monitoring projects.